story/QUAD-95 story/QUAD-96 Refactors the code as per the comments
